{"schema":"decodex/commit/1","summary":"Move scroll capture state access","authority":"manual"}
